requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
dj-stripe==2.8.1
Django==4.2.4
django-cors-headers==4.2.0
django-jsonfield==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
psycopg2==2.9.7
pycparser==2.21
PyJWT==2.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
